Add team mekansm purchase mechanism, and fix many issues. (#61)
Fix chen using persuasion at allies; Fix crystal_maiden cancel frostbite;
Fix the issue that bots don't use abilities when their mana is not enough without mana cost reduction;
Bots now buy mekansm in terms of the team;
Fix ember_spirit rarely use abilities; Fix bounty_hunter don't use track;
Add bristleback hairball; Fix bristleback use quill_spray at unit targets
Fix alchemist berserkers_rage;
Add cm_stop_freezing_field, lycan_wolf_bite;
Stop using sven_strength_of_god, lycan_shapeshift when the target is too far away;
Increase the desire of jakiro using liquid_fire at towers;
Fix doom using devour at ancient targets;
Fix earth_spirit using boulding at unit targets;
Fix furion sprout bugs;
Cancel qop_sonic_wave losing argets after chasing;
Add a feature that bots avoid allies with winters_curse;
Notify teammates when buying team items;
Fix earth_spirit ability usage wrong target;
Fix abyssal_underlord use abilities at creeps;
Fix doom acquired abilities;
Fix ember_spirit abilities;
Decrease the desire of enigma using midnight_pulse at creeps;
Add faceless_void_time_walk_reverse;
Improve natures_prophet not blocking friend melee heroes;
Fix medusa bugs;
Add terrorblade demon_zeal, terror_wave;
Fix windrunner ability level up;
Avoid batrider using napalm when unnecessary;
Fix chen, doom using holy_persuasion or devour at high-level creeps;
Fix clinkz using death_pact at ancient creeps;
Fix earth_spirit, faceless_void using moving abilities at game start;
Low the desire enigma, jakiro using AOE at creeps;
Improve furion sprout to not block friends;
Avoid mirana using ultimate when enemies have dust or gem;
Add ogre_magi_smash;
Fix wraith_king summoning skeleton warriors;
Lower the desire of batrider, troll_warlord, treant using ultimate;
Add troll_warlord_rampage;
Avoid lion, spectre using ultimate when unncessary;
Force every bot to buy dust when enemy has invisibility abilities or items;

+++ ability_item_usage_warlock.lua
@@ -75,7 +75,9 @@ local cast={} cast.Desire={} cast.Target={} cast.Type={}
 local Consider ={}
 local CanCast={utility.NCanCast,function(t)
 	return t:GetTeam() == npcBot:GetTeam() and AbilityExtensions:AllyCanCast(t) and not t:HasModifier("modifier_ice_blast") or AbilityExtensions:NormalCanCast(t) 
-end,utility.NCanCast,utility.UCanCast}
+end,utility.NCanCast,function(t)
+	return AbilityExtensions:NormalCanCast(t, false, DAMAGE_TYPE_MAGICAL, true)
+end}
 local enemyDisabled=utility.enemyDisabled
 
 function GetComboDamage()
@@ -484,13 +486,9 @@ Consider[4]=function()
 	--------------------------------------
 	-- Global high-priorty usage
 	--------------------------------------
-	-- Check for a channeling enemy
-	for _,npcEnemy in pairs( enemys )
-	do
-		if npcEnemy:IsChanneling() and not npcEnemy:IsInvulnerable()
-		then
-			return BOT_ACTION_DESIRE_HIGH, npcEnemy:GetLocation()
-		end
+	local channelling = enemys:First(function(t) return AbilityExtensions:IsChannelingBreakWorthAbility(t, "moderate") and CanCast[4](t) end)
+	if channelling then
+		return BOT_ACTION_DESIRE_MODERATE-0.1, channelling:GetLocation()
 	end
 
 	--try to kill enemy hero
@@ -552,7 +550,6 @@ function AbilityUsageThink()
                 local enemies = npcBot:GetNearbyHeroes(1500, true, BOT_MODE_NONE)
                 enemies = AbilityExtensions:Count(enemies, function(t)
                     return t:HasModifier("modifier_warlock_upheavel")
-                            or GetUnitToLocationDistance(t, upheavelLocation) <= upheavelRadius and AbilityExtensions:GetMagicImmuneRemainingDuration(t) <= 1
                 end)
                 if enemies == 0 then
                     if DotaTime() > upheavelTimer + 1.5 then
